+++ BOCCHI.Common/Config/IConfiguration.cs
@@ -21,5 +21,7 @@ public interface IConfiguration : IPluginConfiguration
 
     TreasureConfig TreasureConfig { get; set; }
 
+    ForkedTowerConfig ForkedTowerConfig { get; set; }
+
     ShoppingConfig ShoppingConfig { get; set; }
 }