+++ BOCCHI.Common/Data/OccultObjectType.cs
@@ -4,4 +4,8 @@ namespace BOCCHI.Common.Data;
 public static class OccultObjectType
 {
     public const uint Carrot = 2010139;
+
+    public const uint Trap = 2014584;
+
+    public const uint BigTrap = 2014585;
 }